Week 14+15 Thomsen_Compton_Buttendorf_Final/requirements.txt
deepface
flask[async]
filetype
Pillow
toml
py-bcrypt